pyproject.toml
[project]
name = "omnipresent"
version = "2.0.0"
description = "Modern YouTube video downloader with cookies auto-extraction"
authors = [
    {name = "Prime_evolution", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["youtube", "downloader", "yt-dlp", "video", "audio"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "PySide6>=6.5.0",
    "selenium>=4.15.0",
    "requests>=2.31.0",
    "websocket-client>=1.6.0",
    "psutil>=5.9.0",
    "browser-cookie3>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/omnipresent"
Repository = "https://github.com/yourusername/omnipresent"
"Bug Tracker" = "https://github.com/yourusername/omnipresent/issues"

[project.scripts]
omnipresent = "main:main"

# ===== Build System =====
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["omnipresent"]

[tool.setuptools.package-data]
omnipresent = ["*.ico", "*.json"]

# ===== PyInstaller конфигурация =====
[tool.pyinstaller]
# Используется для команды: pyinstaller omnipresent.spec

# ===== Briefcase (рекомендуется для GUI) =====
[tool.briefcase]
project_name = "Omnipresent"
bundle = "com.primeevolution"
version = "2.0.0"
url = "https://github.com/yourusername/omnipresent"
license = "MIT license"
author = "Prime_evolution"
author_email = "develop@eclips-team.ru"

[tool.briefcase.app.omnipresent]
formal_name = "Omnipresent"
description = "Modern YouTube video downloader"
icon = "ic"
sources = [
    "main.py",
    "config.py",
    "downloader.py",
    "worker.py",
    "utils.py",
    "browser_cookies.py",
    "ui_qt_widgets.py",
]
requires = [
    "PySide6>=6.5.0",
    "selenium>=4.15.0",
    "requests>=2.31.0",
    "websocket-client>=1.6.0",
    "psutil>=5.9.0",
    "browser-cookie3>=0.19.0",
]

[tool.briefcase.app.omnipresent.windows]
requires = []

# ===== Black (форматирование) =====
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# ===== Pytest =====
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

# ===== MyPy =====
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false